fixed problem with haircolor dye

+++ OpenNos.GameObject/Item/MagicalItem.cs
@@ -66,11 +66,11 @@ namespace OpenNos.GameObject
                         if (EffectValue == 99)
                         {
                             byte nextValue = (byte)random.Next(0, 127);
-                            Session.Character.HairColor = Enum.IsDefined(typeof(HairColorType), nextValue) ? (HairColorType)nextValue : 0;
+                            Session.Character.HairColor = Enum.IsDefined(typeof(HairColorType), (byte)nextValue) ? (HairColorType)nextValue : 0;
                         }
                         else
                         {
-                            Session.Character.HairColor = Enum.IsDefined(typeof(HairColorType), EffectValue) ? (HairColorType)EffectValue : 0;
+                            Session.Character.HairColor = Enum.IsDefined(typeof(HairColorType), (byte)EffectValue) ? (HairColorType)EffectValue : 0;
                         }
                         Session.SendPacket(Session.Character.GenerateEq());
                         Session.CurrentMap?.Broadcast(Session, Session.Character.GenerateIn(), ReceiverType.All);
@@ -98,7 +98,7 @@ namespace OpenNos.GameObject
                         else
                         {
                             // TODO better handling of waxes combined with hairstyle
-                            Session.Character.HairStyle = Enum.IsDefined(typeof(HairStyleType), EffectValue) ? (HairStyleType)EffectValue : 0;
+                            Session.Character.HairStyle = Enum.IsDefined(typeof(HairStyleType), (byte)EffectValue) ? (HairStyleType)EffectValue : 0;
                             Session.SendPacket(Session.Character.GenerateEq());
                             Session.CurrentMap?.Broadcast(Session, Session.Character.GenerateIn(), ReceiverType.All);
                             Inv.Amount--;
